Refactor: Simplify code by removing Python <3.10 compatibility
Following the decision to drop support for Python versions below 3.10,
this commit removes compatibility shims and simplifies the codebase.

Changes include:
- Removed custom fallbacks for `ModuleNotFoundError` and
  `FileNotFoundError` as they are built-in in Python 3.10+.
- Moved the `_create_module` helper function from `magicalimport/compat.py`
  directly into `magicalimport/__init__.py`.
- Updated `magicalimport/__init__.py` to use `importlib.import_module`
  directly.
- Deleted the `magicalimport/compat.py` file as it is no longer needed.
- Updated tests to use built-in `ModuleNotFoundError`.

All tests continue to pass after these changes.
